:sparkles: bug fix

+++ src/main/java/com/pos/features/super_admin/inventory/model/entity/Inventory.java
@@ -34,7 +34,7 @@ public class Inventory implements Serializable {
     private MenuItem menuItem;
 
     @Column(nullable = false)
-    private double quantity= 0.0;
+    private Integer quantity= 0;
 
     @Enumerated(EnumType.STRING)
     private Uom uom;
